DISPATCH-2029: Added a base class and moved some of the http1 adaptor tests to this base class. Also introduced a new test that executes http1 tests over tcp
